refactor test objects and header items
