+++ DragonFruit.Common.Data.Tests/Header/Objects/EchoRequest.cs
@@ -1,7 +1,7 @@
 ﻿// DragonFruit.Common Copyright 2020 DragonFruit Network
 // Licensed under the MIT License. Please refer to the LICENSE file at the root of this project for details
 
-namespace DragonFruit.Common.Data.Tests.Header
+namespace DragonFruit.Common.Data.Tests.Header.Objects
 {
     public class EchoRequest : ApiRequest
     {
