+++ DragonFruit.Common.Data/Headers/HeaderChange.cs
@@ -1,11 +1,11 @@
 ﻿﻿// DragonFruit.Common Copyright 2020 DragonFruit Network
 // Licensed under the MIT License. Please refer to the LICENSE file at the root of this project for details
 
-namespace DragonFruit.Common.Data.Utils
+namespace DragonFruit.Common.Data.Headers
 {
-    internal class ApiClientHeaderChange
+    internal class HeaderChange
     {
-        public ApiClientHeaderChange(string key, string value)
+        public HeaderChange(string key, string value)
         {
             Key = key;
             Value = value;
